+++ tests/smc/test_smc.py
@@ -10,8 +10,7 @@ from absl.testing import absltest
 
 import blackjax
 import blackjax.smc.resampling as resampling
-from blackjax.smc.base import extend_params, init, step
-from blackjax.smc.tempered import update_and_take_last
+from blackjax.smc.base import extend_params, init, step, update_and_take_last
 from blackjax.smc.waste_free import update_waste_free
